+++ connconvert/src/main/java/org/janelia/flyem/connconvert/model2/Neuron.java
@@ -15,7 +15,7 @@ import org.janelia.flyem.connconvert.json.JsonUtils;
 public class Neuron {
 
     @SerializedName("Id")
-    private final Integer id;
+    private final Long id;
 
     @SerializedName("Status")
     private final String status;
@@ -29,7 +29,7 @@ public class Neuron {
     @SerializedName("Size")
     private final Integer size;
 
-    public Neuron(final Integer id,
+    public Neuron(final Long id,
                   final String status,
                   final String name,
                   final String neuronType,
@@ -41,7 +41,7 @@ public class Neuron {
         this.size = size;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
